requirements.txt
absl-py==0.7.0
aiohttp==3.6.2
astor==0.7.1
async-timeout==3.0.1
attrs==19.3.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.1
click==7.1.2
colorama==0.4.3
compare-mt==0.2.5
cycler==0.10.0
decorator==4.3.2
Deprecated==1.2.6
exrex==0.10.5
filelock==3.0.12
fileutils==0.2.2
gast==0.2.2
google==2.0.3
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.18.0
h5py==2.9.0
idna==2.8
idna-ssl==1.1.0
importlib-metadata==1.6.0
jaraco.functools==2.0
javalang==0.12.0
joblib==0.14.1
jsonschema==3.2.0
Keras==2.3.1
kiwisolver==1.1.0
lxml==4.3.3
Markdown==3.0.1
matplotlib==3.1.1
mock==2.0.0
more-itertools==7.0.0
msgpack==0.6.2
multidict==4.7.6
mysql==0.0.2
mysqlclient==1.4.2.post1
networkx==2.2
nltk==3.4
numpy==1.16.1
oandapyV20==0.6.3
pandas==1.0.3
pbr==5.1.2
pkg-resources==0.0.0
portalocker==1.5.0
portend==2.4
protobuf==3.12.0
py-rouge==1.1
py-spy==0.3.3
pydot==1.4.1
PyGithub==1.43.8
PyJWT==1.7.1
pyparsing==2.4.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
pyzmq==19.0.1
ray==0.8.5
redis==3.4.1
repoze.lru==0.7
requests==2.22.0
Routes==2.4.1
sacrebleu==1.3.7
scikit-learn==0.22.1
scipy==1.2.1
scoop==0.7.1.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.8
subword-nmt==0.3.6
tempora==1.14.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.31.1
typing==3.7.4
typing-extensions==3.7.4.2
urllib3==1.25.3
Werkzeug==0.14.1
wrapt==1.11.2
yarl==1.4.2
zc.lockfile==1.4
zipp==3.1.0
